temporary setup command

+++ src/content/list/Commands.js
@@ -84,7 +84,63 @@ class Commands{
 		this.help.params[0] = new globalThis.Command.CommandParam("command", true, []);
 		
 		this.setup = new globalThis.Command("setup", (msg, param, client) => {
+			let key = param[0];
+			let val = param[1];
 			
+			const findSetup = data => {
+				if(typeof(data.servers) === "undefined") data.servers = [{
+					id: msg.guild.id,
+					setup: {}
+				}];
+				
+				let tgtServer = null;
+				
+				data.servers.forEach(server => {
+					if(server.id == msg.guild.id) tgtServer = server;
+				});
+				
+				if(tgtServer == null){
+					let server = {
+						id: msg.guild.id,
+						setup: {}
+					};
+					
+					data.servers.push(server);
+					
+					return server.setup;
+				}else{
+					return tgtServer.setup;
+				};
+			};
+			
+			fs.readFile("setup.json", "utf8", (e, data) => {
+				if(e){
+					console.error(e);
+					msg.reply("error while trying to read setup data.");
+				}else{
+					data = JSON.parse(data);
+					let setup = findSetup(data);
+					
+					if(typeof(val) === "undefined"){
+						if(typeof(setup[key]) === "undefined"){
+							msg.reply("\"" + key + "\" in this server hasn\'t been defined yet.");
+						}else{
+							msg.reply("\"" + key + "\" in this server is currently \"" + setup[key] + "\".");
+						};
+					}else{
+						setup[key] = val;
+						
+						fs.writeFile("setup.json", JSON.stringify(data), "utf8", e => {
+							if(e){
+								console.error(e);
+								msg.reply("error while trying to write setup data.");
+							}else{
+								msg.reply("\"" + key + "\" in this server is now \"" + val + "\".");
+							};
+						})
+					};
+				};
+			});
 		});
 		this.setup.adminOnly = true;
 		this.setup.description = "Configures things for this server.";
@@ -108,8 +164,10 @@ class Commands{
 				};
 			};
 			
-			msg.delete();
-			msg.channel.send(result);
+			if(result.length > 0){
+				msg.delete();
+				msg.channel.send(result);
+			};
 		});
 		this.say.adminOnly = true;
 		this.say.description = "I will say the exact same thing.";
